Update Unity version and package dependencies
The Unity editor version was updated to 2022.3.5f1 from 2022.3.1f1 for compatibility with the latest UNITY API changes and bug fixes. In addition, the versions for the Rider IDE integration and the code coverage tool were updated to their latest versions (3.0.24 and 1.2.4, respectively) for bug fixes and performance improvements. Ensuring that we are using the latest versions of our tools can lead to better project performance and maintainability.