Update InspectorTitlebar to ignore ScriptableObjects
This change modifies the Postfix_ObjectNames_GetInspectorTitle method within the InspectorTitlebar.cs file in the Pillsgood.Unity.Extensions package. Specifically, the method now includes a conditional statement that immediately returns if the 'obj' variable is an instance of a ScriptableObject. This amendment was necessary to avoid incorrectly altering the title of ScriptableObjects in the Unity inspector.