Add script compilation on Inspector Titlebar enable
A code change has been made in the InspectorTitlebar.cs file, particularly to the 'Enabled' setting. A function named OnEnabled has been added to trigger script compilation upon enabling the 'Inspector Titlebar'. This was introduced to ensure code is latest and updated when the title bar is enabled, enhancing the editor's reactiveness.